+++ src/remote_store/aio/_async_azure.py
@@ -39,7 +39,7 @@ if TYPE_CHECKING:
 
 T = TypeVar("T")
 
-_ALL_CAPABILITIES = CapabilitySet(set(Capability) - {Capability.SEEKABLE_READ})
+_ALL_CAPABILITIES = CapabilitySet(set(Capability) - {Capability.SEEKABLE_READ, Capability.ATOMIC_MOVE})
 
 log = logging.getLogger(__name__)